changed price datatype back to Integer

+++ src/main/java/com/fullstack/springreact/entity/Product.java
@@ -19,7 +19,7 @@ public class Product {
 	private String productType;
 	
 
-	private double price;
+	private Integer price;
 	
 
 	private String description;
@@ -48,12 +48,12 @@ public class Product {
 	}
 
 
-	public double getPrice() {
+	public Integer getPrice() {
 		return price;
 	}
 
 
-	public void setPrice(double price) {
+	public void setPrice(Integer price) {
 		this.price = price;
 	}
 
@@ -83,7 +83,7 @@ public class Product {
 	}
 
 
-	public Product(Integer productId, String productType, double price, String description, Integer cashBack) {
+	public Product(Integer productId, String productType, Integer price, String description, Integer cashBack) {
 		super();
 		this.productId = productId;
 		this.productType = productType;
